boilerplate/backend/requirements.txt
# ============================================================================
# VOICE BOT DEPENDENCIES
# ============================================================================

# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Pipecat Voice AI Framework
pipecat-ai==0.0.93

# Daily.co WebRTC Integration
daily-python==0.10.1

# Speech Services
# Deepgram SDK (for STT)
deepgram-sdk==3.8.3

# Azure OpenAI (for LLM)
openai==1.54.5

# ElevenLabs SDK (for TTS)
elevenlabs==1.11.0

# Database
sqlmodel==0.0.22
alembic==1.13.3
psycopg2-binary==2.9.10
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.1
pydantic-settings==2.6.1

# HTTP Client
httpx==0.27.2
aiohttp==3.11.7

# Utilities
python-dotenv==1.0.1
loguru==0.7.3

# Date/Time
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2  # For testing endpoints

# Development Tools
black==24.10.0
ruff==0.7.4
mypy==1.13.0

# ============================================================================
# NOTES
# ============================================================================
# Install with: pip install -r requirements.txt
#
# For development:
# pip install -r requirements.txt
#
# For production (with exact versions):
# pip freeze > requirements-lock.txt
# pip install -r requirements-lock.txt
